Split Extra Room Env Into Separate File
- Moved ExtraItemsDiscreteRoomEnvironment into its own file.
- Reorganised discrete rooms/__init__.py imports.
- Applied Black formatter.
