+++ simpleenvs/envs/discrete_rooms/rooms.py
@@ -1,7 +1,6 @@
 import copy
 import random
 import numpy as np
-from itertools import product
 
 from simpleoptions import TransitionMatrixBaseEnvironment
 
@@ -17,9 +16,7 @@ class DiscreteRoomEnvironment(TransitionMatrixBaseEnvironment):
     Class representing a discrete "rooms-like" gridworld, as is commonly seen in the HRL literature.
     """
 
-    def __init__(
-        self, room_template_file_path, movement_penalty=-0.001, goal_reward=1.0
-    ):
+    def __init__(self, room_template_file_path, movement_penalty=-0.001, goal_reward=1.0):
         """
         Initialises a new DiscreteRoomEnvironment object.
 
@@ -57,10 +54,8 @@ class DiscreteRoomEnvironment(TransitionMatrixBaseEnvironment):
         for y in range(self.gridworld.shape[0]):
             for x in range(self.gridworld.shape[1]):
                 if self.gridworld[y, x] not in CELL_TYPES_DICT:
-                    if not self.gridworld[y,x].replace('-', '', 1).isnumeric():
-                        raise ValueError(
-                            f"Invalid cell type '{self.gridworld[y, x]}' in room template file."
-                        )
+                    if not self.gridworld[y, x].replace("-", "", 1).isnumeric():
+                        raise ValueError(f"Invalid cell type '{self.gridworld[y, x]}' in room template file.")
                 elif CELL_TYPES_DICT[self.gridworld[y, x]] == "start":
                     self.initial_states.append((y, x))
                 elif CELL_TYPES_DICT[self.gridworld[y, x]] == "goal":
@@ -71,12 +66,9 @@ class DiscreteRoomEnvironment(TransitionMatrixBaseEnvironment):
         self.state_space = set()
         for y in range(self.gridworld.shape[0]):
             for x in range(self.gridworld.shape[1]):
-                if (
-                    self.gridworld[y, x] in CELL_TYPES_DICT
-                    and CELL_TYPES_DICT[self.gridworld[y, x]] != "wall"
-                ):
+                if self.gridworld[y, x] in CELL_TYPES_DICT and CELL_TYPES_DICT[self.gridworld[y, x]] != "wall":
                     self.state_space.add((y, x))
-                elif self.gridworld[y,x].replace('-', '', 1).isnumeric():
+                elif self.gridworld[y, x].replace("-", "", 1).isnumeric():
                     self.state_space.add((y, x))
 
     def reset(self, state=None):
@@ -104,9 +96,7 @@ class DiscreteRoomEnvironment(TransitionMatrixBaseEnvironment):
 
     def step(self, action, state=None):
         if state is None:
-            next_state, reward, terminal, info = super().step(
-                action, state=self.current_state
-            )
+            next_state, reward, terminal, info = super().step(action, state=self.current_state)
         else:
             next_state, reward, terminal, info = super().step(action, state=state)
 
@@ -232,8 +222,7 @@ class DiscreteRoomEnvironment(TransitionMatrixBaseEnvironment):
             # if next state is a wall return to the current state
             if (
                 self.gridworld[next_state[0]][next_state[1]] in CELL_TYPES_DICT
-                and CELL_TYPES_DICT[self.gridworld[next_state[0]][next_state[1]]]
-                == "wall"
+                and CELL_TYPES_DICT[self.gridworld[next_state[0]][next_state[1]]] == "wall"
             ):
                 next_state = (state[0], state[1])
 
@@ -242,135 +231,11 @@ class DiscreteRoomEnvironment(TransitionMatrixBaseEnvironment):
             else:
                 if (
                     self.gridworld[next_state[0]][next_state[1]] not in CELL_TYPES_DICT
-                    and self.gridworld[next_state[0]][next_state[1]].replace('-', '', 1).isnumeric()
+                    and self.gridworld[next_state[0]][next_state[1]].replace("-", "", 1).isnumeric()
                 ):
                     reward = float(self.gridworld[next_state[0]][next_state[1]]) + self.movement_penalty
                 else:
                     reward = self.movement_penalty
-
-            successor_states.append(((next_state, reward), 1.0 / len(actions)))
-
-        return successor_states
-
-
-class ExtraItemsDiscreteRoomEnvironment(DiscreteRoomEnvironment):
-    def __init__(
-        self, room_template_file_path, movement_penalty=-0.001, goal_reward=1.0
-    ):
-        self.basic_init = True
-        super().__init__(room_template_file_path, movement_penalty, goal_reward)
-        self.basic_init = False
-        self.positions = self.state_space
-        self.item_locations = self.get_item_locations()
-        self.state_space, self.terminal_states = self.adjust_for_item(
-            self.state_space, self.terminal_states, self.item_locations
-        )
-        self.transition_matrix = self._compute_transition_matrix()
-
-    def get_item_locations(self):
-        return [
-            (x, y)
-            for x in range(self.gridworld.shape[0])
-            for y in range(self.gridworld.shape[1])
-            if self.gridworld[x, y].replace('-', '', 1).isnumeric()
-        ]
-
-    def adjust_for_item(self, state_space, terminal_states, item_locations):
-        """
-        Given the number of item positions, adjusts the state space and terminal states.
-        Increases the state space to include all possible combinations of item collection states.
-        Removes the original item positions from the state space.
-        Adds all combinations of the terminal states with different amounts of item.
-        """
-        # Create a list of all possible combinations of item collection states
-        num_items = len(item_locations)
-        item_combinations = list(product([0, 1], repeat=num_items))
-
-        # Create the modified states array
-        modified_states = []
-
-        for state in state_space:
-            for combination in item_combinations:
-                modified_state = list(state) + list(combination)
-                while modified_state and modified_state[-1] == 0:
-                    modified_state.pop()
-                trimmed_state = tuple(modified_state)
-                modified_states.append(trimmed_state)
-                if state in terminal_states and trimmed_state not in terminal_states:
-                    terminal_states.append(trimmed_state)
-
-        # remove the original item positions from the state space
-        for item in item_locations:
-            modified_states.remove(item)
-
-        return modified_states, terminal_states
-
-    def has_picked_up_item(self, state):
-        position = (state[0], state[1])
-        if len(state) <= 2 or not hasattr(self, "item_locations"):
-            return False
-        # get the index of the x,y position in the item locations
-        item_index = self.item_locations.index(position)
-        if len(state) <= 2 + item_index:
-            # print(f"ERR:\tstate: {state}\titem_index: {item_index}")
-            return False
-        return state[2 + item_index] == 1
-
-    def get_successors(self, state=None, actions=None):
-        if (
-            self.basic_init
-        ):  # if initial setup is being established, use the basic setup
-            return super().get_successors(state=state, actions=actions)
-        if state is None:
-            state = self.current_state
-
-        if actions is None:
-            actions = self.get_available_actions(state=state)
-
-        successor_states = []
-        for action in actions:
-            next_state = copy.deepcopy(state)
-            if ACTIONS_DICT[action] == "DOWN":
-                next_state = (state[0] + 1, state[1], *state[2:])
-            elif ACTIONS_DICT[action] == "UP":
-                next_state = (state[0] - 1, state[1], *state[2:])
-            elif ACTIONS_DICT[action] == "RIGHT":
-                next_state = (state[0], state[1] + 1, *state[2:])
-            elif ACTIONS_DICT[action] == "LEFT":
-                next_state = (state[0], state[1] - 1, *state[2:])
-            # if next state is a wall return to the current state
-            if (
-                self.gridworld[next_state[0]][next_state[1]] in CELL_TYPES_DICT
-                and CELL_TYPES_DICT[self.gridworld[next_state[0]][next_state[1]]]
-                == "wall"
-            ):
-                next_state = (state[0], state[1], *state[2:])
-
-            if self.is_state_terminal(state=(next_state[0], next_state[1])):
-                reward = self.goal_reward
-            else:  # state is either a floor or a item position
-                # if ns is a item position and the item has not been picked up
-                if (
-                    self.gridworld[next_state[0]][next_state[1]] not in CELL_TYPES_DICT
-                    and self.gridworld[next_state[0]][next_state[1]].replace('-', '', 1).isnumeric()
-                    and not self.has_picked_up_item(next_state)
-                ):
-                    # get the reward at that position
-                    reward = (
-                        float(self.gridworld[next_state[0]][next_state[1]])
-                        + self.movement_penalty
-                    )
-                    # get the id of which item it is
-                    item_index = self.item_locations.index(next_state[:2])
-                    next_state = list(next_state)
-                    # fill out the remainder of the state item flags
-                    while len(next_state) < 2 + item_index + 1:
-                        next_state.append(0)
-                    next_state[2 + item_index] = 1
-                    next_state = tuple(next_state)
-                else:
-                    reward = self.movement_penalty
-
 
             successor_states.append(((next_state, reward), 1.0 / len(actions)))
 
@@ -432,18 +297,6 @@ with pkg_resources.path(data, "ramesh_maze.txt") as path:
 
 with pkg_resources.path(data, "wide_path.txt") as path:
     wide_path = path
-
-with pkg_resources.path(data, "basic_reward_room.txt") as path:
-    basic_reward_room = path
-
-with pkg_resources.path(data, "double_reward_room.txt") as path:
-    double_reward_room = path
-
-with pkg_resources.path(data, "basic_penalty_room.txt") as path:
-    basic_penalty_room = path
-
-with pkg_resources.path(data, "double_penalty_room.txt") as path:
-    double_penalty_room = path
 
 
 class DiscreteDefaultTwoRooms(DiscreteRoomEnvironment):
@@ -624,53 +477,3 @@ class WidePath(DiscreteRoomEnvironment):
 
     def __init__(self, movement_penalty=-0.001, goal_reward=1):
         super().__init__(wide_path, movement_penalty, goal_reward)
-
-
-class BasicRewardRoom(ExtraItemsDiscreteRoomEnvironment):
-    """
-    An 11x11 grid with start in the top left and goal in the bottom right.
-    The environment has an additional reward of 10 in the bottom left.
-    Goal Reward: +1
-    Movement Penalty: -0.01
-    """
-
-    def __init__(self, movement_penalty=-0.001, goal_reward=1):
-        super().__init__(basic_reward_room, movement_penalty, goal_reward)
-
-
-class DoubleRewardRoom(ExtraItemsDiscreteRoomEnvironment):
-    """
-    An 11x11 grid with start in the top left and goal in the bottom right.
-    The environment has two additional rewards of 10 in the bottom left and top right.
-    Goal Reward: +1
-    Movement Penalty: -0.01
-    """
-
-    def __init__(self, movement_penalty=-0.001, goal_reward=1):
-        super().__init__(double_reward_room, movement_penalty, goal_reward)
-
-
-class BasicPenaltyRoom(ExtraItemsDiscreteRoomEnvironment):
-    """
-    An 11x11 grid with start in the top left and goal in the bottom right.
-    The environment has an additional penalties of -10 in the bottom left that can be picked up.
-    Once a penalty state has been visited, returning to this position will not result in an additional penalty.
-    Goal Reward: +1
-    Movement Penalty: -0.01
-    """
-
-    def __init__(self, movement_penalty=-0.001, goal_reward=1):
-        super().__init__(basic_penalty_room, movement_penalty, goal_reward)
-
-
-class DoublePenaltyRoom(ExtraItemsDiscreteRoomEnvironment):
-    """
-    An 11x11 grid with start in the top left and goal in the bottom right.
-    The environment has two additional penalties of -10 in the bottom left and top right that can be picked up.
-    Once a penalty state has been visited, returning to this position will not result in an additional penalty.
-    Goal Reward: +1
-    Movement Penalty: -0.01
-    """
-
-    def __init__(self, movement_penalty=-0.001, goal_reward=1):
-        super().__init__(double_penalty_room, movement_penalty, goal_reward)
